==> scripts/01_mixcr_qc.Rmd <==
---
title: "01 - MiXCR quality control"
author: "Teresa Rubio"
date: "February, 2021"
output:
    html_document:
      code_folding: hide
      self_contained: true
      keep_md: true
params:
    workDir: !r getwd()
    outputDir: !r getwd()
    sampleInfo: !r NULL
---

```{r, setup, include=FALSE}
knitr::opts_knit$set(root.dir = params$workDir, echo=TRUE, dev=c('png','pdf'))
```

# Packages
```{r, message=FALSE}
library(ggplot2)
library(plyr)
library(RColorBrewer)
library(tibble)
library(tidyr)
```

# Work enviroment
```{r}
# Define input/output
sampleInfoFile <- params$sampleInfo
dir.create(params$outputDir, recursive=T)

# Load sample info
SampleInfo <- read.table(sampleInfoFile, sep = ",", header = T, check.names = F)
levels <- unique(SampleInfo$Group)
SampleInfo$Group <- factor(SampleInfo$Group, levels = levels)

# Color palette for sample groups
cbPalette <- c("#56B4E9","#CB181D","#000000","#CC79A7","#db6d00","#009E73",
               "#006ddb","#b66dff","#924900","#490092","#999999","#F0E442")
```


# Alignment Report
```{r, include=FALSE}
# Alignment files (report00) in SampleInfo order
files <- list.files(path=params$workDir, pattern="*report00")
files <- files[match(SampleInfo$SampleID, gsub(".report00", "", files))]

# Concatenate all repertoire reports in one
dfs <- lapply(files, read.table, row.names = 1, sep=":", comment.char="=", skip = 6)
dfs2 <- lapply(lapply(dfs, t), as.data.frame)
alignment_reports <- t(plyr::join_all(dfs = dfs2, type = "full"))
colnames(alignment_reports) <- SampleInfo$SampleID

# Remove (%) values
alignment_reports <- apply(alignment_reports, 1, stringr::str_remove, pattern = " [(\\)].+")
# Change to numeric
alignment_reports <- apply(alignment_reports, 2, as.numeric)
# Add rownames
rownames(alignment_reports) <- SampleInfo$SampleID

# Create alignment report table
alignment_reports <- data.frame(alignment_reports, check.names = F)
alignment_reports <- cbind(SampleID = rownames(alignment_reports), alignment_reports)

write.table(alignment_reports, file = file.path(params$outputDir, "01_alignment_reports.tsv"),
            sep = "\t", row.names = F, col.names = T, quote = F)
```


## Plots
```{r}
# Rmarkdown plot
df_matrix <- tibble::as_tibble(alignment_reports)
df_matrix$SampleID <- factor(df_matrix$SampleID, levels=df_matrix$SampleID)

df_gather_chains <- tidyr::gather(data = df_matrix[,c(1, grep("^[A-Z]{3} chains", colnames(df_matrix)))],
                                  key = "parameter", value = "reads", -1)
df_gather_chains$parameter <- factor(df_gather_chains$parameter,
                                     levels = unique(df_gather_chains$parameter))
colnames(df_gather_chains)[1] <- "SampleID"

# color
cbPalette4chains <- RColorBrewer::brewer.pal(length(unique(df_gather_chains$parameter)), "Set3")
sample_color <- cbPalette[factor(SampleInfo$Group)]

# Plot                                           
alignment_plot <- ggplot(df_gather_chains,
             aes(fill = parameter,
                 y = reads,
                 x = SampleID)) +
      geom_bar(position="stack",
               stat="identity") +
      ggtitle("MiXCR alignment report") +
      ylab("Number of reads") +
      theme_light() +
      theme(axis.text.x = element_text(angle = 90,
                                       color = sample_color)) +
      scale_fill_manual(values = cbPalette4chains)
alignment_plot
```

## Output files
```{r, results='hide'}
save(alignment_plot, file = file.path(params$workDir, "01_alignment_plot.Rda"))

png(file.path(params$outputDir, "01_alignment_plot.png"), width = 30, height = 15, units = "cm", res = 300)
alignment_plot
dev.off()
```


# Assemble report
```{r, include=FALSE}
# Assemble files (report04) in SampleInfo order
files <- list.files(path=params$workDir, pattern="*report04")
files <- files[match(SampleInfo$SampleID, gsub(".report04", "", files))]

# Concatenate all repertoire reports in one
dfs <- lapply(files, read.table, row.names = 1, sep=":", comment.char="=", skip = 7)
dfs2 <- lapply(lapply(dfs, t), as.data.frame)
assemble_reports <- t(plyr::join_all(dfs = dfs2, type = "full"))
colnames(assemble_reports) <- SampleInfo$SampleID

# Remove (%) values
assemble_reports <- apply(assemble_reports, 1, stringr::str_remove, pattern = " [(\\)].+")
# Change to numeric
assemble_reports <- apply(assemble_reports, 2, as.numeric)
# Add rownames
rownames(assemble_reports) <- SampleInfo$SampleID

# Create alignment report table
assemble_reports <- data.frame(assemble_reports, check.names = F)
assemble_reports <- cbind(SampleID = rownames(assemble_reports), assemble_reports)

write.table(assemble_reports, file = file.path(params$outputDir, "01_assemble_reports.tsv"),
            sep = "\t", row.names = F, col.names = T, quote = F)
```

## Plots
```{r}
# Rmarkdown plot
df_matrix_assemble <- data.frame(assemble_reports, check.names = F)
df_matrix_assemble <- tibble::as_tibble(df_matrix_assemble)
df_matrix_assemble$SampleID <- factor(df_matrix_assemble$SampleID, levels=df_matrix_assemble$SampleID)

df_gather_chains <- tidyr::gather(data = df_matrix_assemble[,c(1, grep("^[A-Z]{3} chains", colnames(df_matrix_assemble)))], key = "parameter", value = "reads", -1)
df_gather_chains$parameter <- factor(df_gather_chains$parameter,
                                     levels = unique(df_gather_chains$parameter))
# color
cbPalette4chains <- RColorBrewer::brewer.pal(length(unique(df_gather_chains$parameter)), "Set3")
sample_color <- cbPalette[factor(SampleInfo$Group)]

# Plot
assemble_plot <- ggplot(df_gather_chains,
             aes(fill = parameter,
                 y = reads,
                 x = SampleID)) +
      geom_bar(position="stack",
               stat="identity") +
      ggtitle("MiXCR assemble report") +
      ylab("Number of clonotypes") +
      theme_light() +
      theme(axis.text.x = element_text(angle = 90,
                                       color = sample_color)) +
      scale_fill_manual(values = cbPalette4chains)
assemble_plot
```

## Output files
```{r, results='hide'}
save(assemble_plot, file = file.path(params$workDir, "01_assemble_plot.Rda"))

png(file.path(params$outputDir, "01_assemble_plot.png"), width = 30, height = 15, units = "cm", res = 300)
assemble_plot
dev.off()
```
