==> scripts/02_datafiltering.Rmd <==
---
title: "02 - Data Filtering and Evenness calculation"
author: "Teresa Rubio"
date: "February, 2021"
output:
    html_document:
      code_folding: hide
params:
  inputDir: !r getwd()
  workDir: !r getwd()
  outputDir: !r getwd()
  sampleInfo: !r NULL
  sampleLevels: !r NULL
---

# Packages
```{r, message=FALSE}
library(vegan)
library(microbiome)
```

# Work enviroment
```{r}
# Define input/work/output directories
sampleInfoFile <- params$sampleInfo
dir.create(params$outputDir, recursive=T)

# Load sample info
SampleInfo <- read.table(sampleInfoFile, sep = ",", header = T, check.names = F)
levels <- unique(SampleInfo$Group)
SampleInfo$Group <- factor(SampleInfo$Group, levels = levels)
```

# Clones from mixcr output
Clones were included in the analysis if they had a minimal abundance read of 2 read counts (thus avoiding singletones) and their CDR3βs were of 4 amino acids minimum length. Additionally, mixcr output contained BCR+TCR clones and we exclude everything that is not TCR (keeping TRA+TRB+TRG+TRD).

Two types of clones are calculated:

* TRAclones: TRA chain
* TRBclones: TRB chain

```{r}
# List input files and keep SampleInfo order
files <- list.files(path = params$inputDir, pattern="*.clonotypes.ALL.txt")
files <- files[match(SampleInfo$SampleID, gsub(".clonotypes.ALL.txt", "", files))]

# Create a list of clone tables
clonotypes_all <- setNames(lapply(files, read.delim),nm = SampleInfo$SampleID)
```


## TRA clonotypes
```{r}
# Exclude everything that is not TRA
clonotypes_TRA <- lapply(clonotypes_all, function(i){
  i = i[grep("TRA", i$allVHitsWithScore),]
  })

# Summary of filter criteria
summary_table_TRA <- do.call(rbind, lapply(names(clonotypes_TRA), function(i){
  table = clonotypes_TRA[[i]]
  sample = i
  TRAclones = nrow(table)
  CDR3aa_minlength = min(nchar(clonotypes_TRA[[i]]$aaSeqCDR3))
  filterby_readcount = nrow(table[table$cloneCount >= 2,])
  filterby_CDR3aa = nrow(table[nchar(table$aaSeqCDR3) >= 4,])
  filterby_both = nrow(table[table$cloneCount >= 2 & nchar(table$aaSeqCDR3) >= 4,])
  cbind(sample,TRAclones,filterby_readcount,CDR3aa_minlength,filterby_CDR3aa,filterby_both)
  }))

# Filtering and output
clones_TRA_filtered <- lapply(clonotypes_TRA, function(i){
  # Filters
  i = i[i$cloneCount >= 2,]
  i = i[i$aaSeqCDR3 >= 4,]
  # Recalculate clone Fraction
  recalculatedcloneFraction = microbiome::transform(i[,"cloneCount",drop=F], "compositional")
  if (nrow(i)>0){
    colnames(recalculatedcloneFraction) = "recalculatedcloneFraction"
  }
  cbind(i, recalculatedcloneFraction)
})

clones_filtered <- clones_TRA_filtered
save(clones_filtered, file = file.path(params$workDir,
                                           "clones_TRA_filtered.Rda"))
```

## TRB clonotypes
```{r}
# Exclude everything that is not TRB
clonotypes_TRB <- lapply(clonotypes_all, function(i){
  i = i[grep("TRB", i$allVHitsWithScore),]
  })

# Summary of filter criteria
summary_table_TRB <- do.call(rbind, lapply(names(clonotypes_TRB), function(i){
  table = clonotypes_TRB[[i]]
  sample = i
  TRBclones = nrow(table)
  CDR3aa_minlength = min(nchar(clonotypes_TRB[[i]]$aaSeqCDR3))
  filterby_readcount = nrow(table[table$cloneCount >= 2,])
  filterby_CDR3aa = nrow(table[nchar(table$aaSeqCDR3) >= 4,])
  filterby_both = nrow(table[table$cloneCount >= 2 & nchar(table$aaSeqCDR3) >= 4,])
  cbind(sample,TRBclones,filterby_readcount,CDR3aa_minlength,filterby_CDR3aa,filterby_both)
  }))

# Filtering and output
clones_TRB_filtered <- lapply(clonotypes_TRB, function(i){
  # Filters
  i = i[i$cloneCount >= 2,]
  i = i[i$aaSeqCDR3 >= 4,]
  # Recalculate clone Fraction
  recalculatedcloneFraction = microbiome::transform(i[,"cloneCount",drop=F], "compositional")
  colnames(recalculatedcloneFraction) = "recalculatedcloneFraction"
  cbind(i, recalculatedcloneFraction)
})

clones_filtered <- clones_TRB_filtered
save(clones_filtered, file = file.path(params$workDir,
                                           "clones_TRB_filtered.Rda"))
```


# Shannon Evenness (SE)

## TRA - SE
```{r}
TRA_SE <- do.call(rbind, lapply(clones_TRA_filtered, function(x){
  diversity = exp(vegan::renyi(x = x[,"recalculatedcloneFraction"], scales = 1))
  richness = exp(vegan::renyi(x = x[,"recalculatedcloneFraction"], scales = 0))
  SE = diversity/richness
  }))
colnames(TRA_SE) <- "Shannon-Evenness"
```

## TRB - SE
```{r}
TRB_SE <- do.call(rbind, lapply(clones_TRB_filtered, function(x){
  diversity = exp(vegan::renyi(x = x[,"recalculatedcloneFraction"], scales = 1))
  richness = exp(vegan::renyi(x = x[,"recalculatedcloneFraction"], scales = 0))
  SE = diversity/richness
  }))
colnames(TRB_SE) <- "Shannon-Evenness"
```

# Output files
```{r, results='hide', warning=FALSE}
# Summary table
summary_table_TRA <- cbind(summary_table_TRA, TRA_SE)
summary_table_TRB <- cbind(summary_table_TRB, TRB_SE)

write.table(summary_table_TRA, file = file.path(params$outputDir, "02_filtering-evenness_summarytable_TRA.csv"),
            sep = "\t", row.names = F, col.names = T, quote = F)
write.table(summary_table_TRB, file = file.path(params$outputDir, "02_filtering-evenness_summarytable_TRB.csv"),
            sep = "\t", row.names = F, col.names = T, quote = F)

# Filtered files
dir.create("clones_TRA_filtered")
dir.create("clones_TRB_filtered")


lapply(names(clones_TRA_filtered), function(x){
  write.table(clones_TRA_filtered[[x]],
              file = file.path(params$workDir, "clones_TRA_filtered", 
                               paste0(x, ".txt")),
              quote = F, sep = "\t", row.names = F)
})

lapply(names(clones_TRB_filtered), function(x){
  write.table(clones_TRB_filtered[[x]],
              file = file.path(params$workDir, "clones_TRB_filtered", 
                               paste0(x, ".txt")),
              quote = F, sep = "\t", row.names = F)
})
```
